requirements.txt
absl-py==1.4.0
anyio==3.7.1
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==6.0.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.3.1
catkin==0.7.18
catkin-pkg==0.5.2
catkin-pkg-modules==0.5.2
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.0.3
cmake==3.27.0
colorama==0.4.4
comm==0.1.3
command-not-found==0.3
common @ file:///home/rail/diya/RAIL-onboarding-local/modules/common
cryptography==3.4.8
cupshelpers==1.0
cv-bridge==1.16.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.7
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
docutils==0.17.1
duplicity==0.8.21
EasyProcess==1.1
empy==3.3.4
entrypoint2==1.1
environments @ file:///home/rail/diya/RAIL-onboarding-local/modules/environments
example @ file:///home/rail/diya/RAIL-onboarding-local/modules/example
exceptiongroup==1.1.2
executing==1.2.0
fasteners==0.14.1
fastjsonschema==2.18.0
filelock==3.12.2
flake8==4.0.1
fonttools==4.29.1
fqdn==1.5.1
fs==2.4.12
fsspec==2023.6.0
future==0.18.2
gast==0.5.2
genmsg==0.5.16
genpy==0.6.16
gnupg==2.3.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gridmap @ file:///home/rail/diya/RAIL-onboarding-local/modules/gridmap
grpcio==1.56.2
html5lib==1.1
httplib2==0.20.2
huggingface-hub==0.16.4
idna==3.3
imageio==2.31.1
importlib-metadata==4.6.4
iniconfig==1.1.1
ipykernel==6.25.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
lark==1.1.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
learning @ file:///home/rail/diya/RAIL-onboarding-local/modules/learning
lit==16.0.6
llvmlite==0.40.1
lockfile==0.12.2
louis==3.20.0
lsp @ file:///home/rail/diya/RAIL-onboarding-local/modules/lsp
lsp-accel @ file:///home/rail/diya/RAIL-onboarding-local/modules/lsp_accel
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mistune==3.0.1
monotonic==1.6
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==0.0.0
mss==9.0.1
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.1
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.1
notebook==7.0.0
notebook_shim==0.2.3
numba==0.57.1
numpy==1.21.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
olefile==0.46
opencv-python==4.8.0.74
overrides==7.3.1
packaging==21.3
pandas==2.0.3
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.9.1
pluggy==0.13.0
ply==3.11
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybullet==3.2.5
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.18.0
pycups==2.0.1
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyscreenshot==3.1
pytest==7.4.0
pytest-html==3.2.0
pytest-metadata==3.0.0
pytest-timeout==2.1.0
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-json-logger==2.0.7
pythran==0.10.0
pytz==2022.1
PyWavelets==1.4.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.0
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
roman==3.3
rosbag==1.15.11
roscpp==1.15.11
rosdistro-modules==0.9.0
rosgraph==1.15.11
rosgraph-msgs==1.11.3.post2
roslib==1.14.7.post0
roslz4==1.14.3.post2
rospkg==1.5.0
rospkg-modules==1.5.0
rospy==1.15.11
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scipy==1.8.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
Send2Trash==1.8.2
sensor-msgs==1.13.1
Shapely==1.6.4
six==1.16.0
sknw==0.14
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.1
ssh-import-id==5.11
stack-data==0.6.2
std-msgs==0.5.13.post0
sympy==1.9
systemd-python==234
tensorboard==2.13.0
tensorboard-data-server==0.7.1
terminado==0.17.1
tifffile==2023.7.18
timm==0.9.2
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
unitybridge @ file:///home/rail/diya/RAIL-onboarding-local/modules/unitybridge
uri-template==1.3.0
urllib3==1.26.5
usb-creator==0.3.7
vulture==2.7
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
widgetsnbextension==4.0.8
xdg==5
xkit==0.0.0
zipp==1.0.0
